requirements.txt
babbage==0.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
normality==2.4.0
psycopg2==2.9.3
python-dateutil==2.8.2
requests==2.28.1
openpyxl==3.0.10
iatikit==3.3.1
iatiflattener==0.1.5
